backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4,<6.0.0",
    "djangorestframework>=3.16.0,<4.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "django-cors-headers>=4.7.0,<5.0.0",
    "djangorestframework-simplejwt>=5.5.1,<6.0.0",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=12.0.0,<13.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
